# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/wallet/{user_id}": {
            "get": {
                "tags": [
                    "wallet"
                ],
                "summary": "Get wallet by user id",
                "description": "Get wallet details of the user by user id",
                "operationId": "getWalletByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "wallet_id"
                                        },
                                        "user_id": {
                                            "type": "string",
                                            "example": "user_id"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000.5
                                        },
                                        "coin_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000.5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-07-04 14:22:22"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-07-04 14:22:22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Wallet not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}